requirements.txt
aiogram==3.4.1
openai>=1.3.0
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib